[iter14] refactor agent's logger
