+++ internal/agent/agent.go
@@ -11,14 +11,16 @@ import (
 	"math/rand"
 	"net"
 	"net/http"
+	"os"
 	"runtime"
 	"strconv"
 	"time"
 
 	"github.com/ospiem/mcollector/internal/models"
+	"github.com/ospiem/mcollector/internal/tools"
+	"github.com/rs/zerolog"
 
 	"github.com/ospiem/mcollector/internal/agent/config"
-	"github.com/rs/zerolog/log"
 )
 
 const defaultSchema = "http://"
@@ -36,8 +38,9 @@ func Run() error {
 	if err != nil {
 		return fmt.Errorf("run agent error: %w", err)
 	}
-
-	log.Info().Msgf("Start server\nPush to %s\nCollecting metrics every %v\n"+
+	logger := zerolog.New(os.Stderr).With().Timestamp().Logger()
+	tools.SetLogLevel(cfg.LogLevel)
+	logger.Info().Msgf("Start server\nPush to %s\nCollecting metrics every %v\n"+
 		"Send metrics every %v\n", cfg.Endpoint, cfg.PollInterval, cfg.ReportInterval)
 
 	m := runtime.MemStats{}
@@ -56,14 +59,14 @@ func Run() error {
 		case <-mTicker.C:
 			err := GetMetrics(&m, metrics)
 			if err != nil {
-				log.Err(err)
+				logger.Err(err)
 			}
 			pollCounter++
 		case <-reqTicker.C:
 			for name, value := range metrics {
 				v, err := strconv.ParseFloat(value, 64)
 				if err != nil {
-					log.Error().Msg("error convert string to float")
+					logger.Error().Msg("error convert string to float")
 					break
 				}
 				metricSlice = append(metricSlice, models.Metrics{MType: gauge, ID: name, Value: &v})
@@ -76,12 +79,12 @@ func Run() error {
 			sleepTime := 1 * time.Second
 			for {
 				var opError *net.OpError
-				err = doRequestWithJSON(cfg, metricSlice, client)
+				err = doRequestWithJSON(cfg, metricSlice, client, logger)
 				if err == nil {
 					break
 				}
 				if errors.As(err, &opError) || errors.Is(err, errRetryableHTTPStatusCode) {
-					log.Error().Err(err).Msgf("%s, will retry in %v", cannotCreateRequest, sleepTime)
+					logger.Error().Err(err).Msgf("%s, will retry in %v", cannotCreateRequest, sleepTime)
 					time.Sleep(sleepTime)
 					attempt++
 					sleepTime += repeatFactor * time.Second
@@ -90,7 +93,7 @@ func Run() error {
 					}
 					break
 				}
-				log.Error().Err(err).Msgf("cannot do request, failed %d times", retryAttempts)
+				logger.Error().Err(err).Msgf("cannot do request, failed %d times", retryAttempts)
 			}
 			metricSlice = nil
 			pollCounter = 0
@@ -98,7 +101,7 @@ func Run() error {
 	}
 }
 
-func doRequestWithJSON(cfg config.Config, metrics []models.Metrics, client *http.Client) error {
+func doRequestWithJSON(cfg config.Config, metrics []models.Metrics, client *http.Client, l zerolog.Logger) error {
 	const wrapError = "do request error"
 
 	jsonData, err := json.Marshal(metrics)
@@ -125,7 +128,7 @@ func doRequestWithJSON(cfg config.Config, metrics []models.Metrics, client *http
 	request.Header.Set("Content-Type", "application/json")
 	request.Header.Set("Content-Encoding", "gzip")
 	if cfg.Key != "" {
-		request.Header.Set("HashSHA256", generateHash(cfg.Key, jsonData))
+		request.Header.Set("HashSHA256", generateHash(cfg.Key, jsonData, l))
 	}
 
 	r, err := client.Do(request)
@@ -157,11 +160,12 @@ func isStatusCodeRetryable(code int) bool {
 	}
 }
 
-func generateHash(key string, data []byte) string {
+func generateHash(key string, data []byte, l zerolog.Logger) string {
+	logger := l.With().Str("func", "generateHash").Logger()
 	h := hmac.New(sha256.New, []byte(key))
 	_, err := h.Write(data)
 	if err != nil {
-		log.Error().Err(err).Msg("cannot hash data")
+		logger.Error().Err(err).Msg("cannot hash data")
 		return ""
 	}
 	return string(h.Sum(nil))
