+++ internal/agent/config/flag.go
@@ -14,6 +14,7 @@ func ParseFlag(c *Config) {
 	flag.IntVar(&ri, "r", defaultReportInterval, "Configure the agent's report interval")
 	flag.IntVar(&pi, "p", defaultPollInterval, "Configure the agent's poll interval")
 	flag.StringVar(&c.Key, "k", "", "Set key for hash function")
+	flag.StringVar(&c.LogLevel, "l", "info", "Configure the agent's log level")
 	flag.Parse()
 
 	c.ReportInterval = time.Duration(ri) * time.Second
